requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
black==24.3.0
celery==5.3.6
celery-progress==0.3
cfgv==3.4.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
distlib==0.3.8
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-countries==7.5.1
django-debug-toolbar==4.3.0
django-faker==0.2
django-mathfilters==1.0.0
django-modeltranslation==0.18.13b1
django-shell-plus==1.1.7
fake-factory==9999.9.9
Faker==24.2.0
filelock==3.13.3
flake8==7.0.0
gunicorn==21.2.0
identify==2.5.35
isort==5.13.2
kombu==5.3.5
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==24.0
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.2.0
pre-commit==3.7.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
redis==5.0.3
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2024.1
vine==5.1.0
virtualenv==20.25.1
wcwidth==0.2.13
